src/requirements.txt
Django===2.0
psycopg2
django-filter==1.1.0
